requirements.txt
streamlit==1.50.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.34.0
PyPDF2==3.0.1
numpy==1.26.4
